Things are working now!
The two filters applied to  list<BlockPos> removed the inside of the cube created by it. This made it appear that the mod broke the tables wear it was functioning as expected.

Renamed all packages, resources, and mixin classes from better.enchantment to bet.enchanttable and better-enchantment to better-enchantment-tables. Updated mod metadata, mixin configs, and Gradle properties accordingly. Removed old files and added new entrypoint classes and mixins to match the new naming and structure.